Song domain class deletion and doSearch Item in UniversalController
