+++ grails-app/controllers/ivvq/MovieController.groovy
@@ -104,14 +104,4 @@ class MovieController {
         }
     }
 
-/**
- * Searches movies corresponding to a search string passed in params
- * @return a list of movies as search result
- */
-    def doSearchMovies() {
-        params.max = 5
-        def moviesList = movieService.searchMovies(params)
-        render(view: 'searchResults', model: [movieInstanceList:moviesList, movieInstanceCount: moviesList.totalCount])
-
-    }
 }
